Backout 24dbe7da7370:a319694b46d5 (bug 1069556) for build bustage on various platforms on a CLOSED TREE
UltraBlame original commit: 9087dd73457a98ccf303d0e2613135f31fc172c9

+++ toolkit/crashreporter/google-breakpad/src/common/md5.cc
@@ -58,7 +58,7 @@ void MD5Init(struct MD5Context *ctx)
 
 
 
-void MD5Update(struct MD5Context *ctx, unsigned char const *buf, size_t len)
+void MD5Update(struct MD5Context *ctx, unsigned char const *buf, unsigned len)
 {
   u32 t;
